Removed Unused Junit5 dependency and package-lock.json
JUnit 5 is currently part of our dependencies, although JUnit 4 is being
used everywhere. Removing it will prevent conflicts in the future. On
top of that, the version for the junit-jupiter-api was LATEST, which is
deprecated.

Additionally, yarn warns that having a package-lock.json can cause
problems in the future. It was therefore removed